[wasm][cleanup] Stop using base::Vector::length()
Replacing it with Vector::size() where that makes sense.

Bug: 393395932
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7146164
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103679}

+++ src/wasm/baseline/liftoff-compiler.cc
@@ -2015,7 +2015,7 @@ class LiftoffCompiler {
 
   void ResumeHandler(FullDecoder* decoder,
                      base::Vector<const HandlerCase> handlers,
-                     int handler_index, Value* cont) {
+                     size_t handler_index, Value* cont) {
     // "Resume" bails out before this can be reached.
     UNREACHABLE();
   }
