+++ src/wasm/streaming-decoder.h
@@ -42,8 +42,8 @@ class V8_EXPORT_PRIVATE StreamingProcessor {
   // finished successfully and the decoding should continue.
   virtual bool ProcessCodeSectionHeader(int num_functions, uint32_t offset,
                                         std::shared_ptr<WireBytesStorage>,
-                                        int code_section_start,
-                                        int code_section_length) = 0;
+                                        size_t code_section_start,
+                                        size_t code_section_length) = 0;
 
   // Process a function body. Returns true if the processing finished
   // successfully and the decoding should continue.
